Back/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning
scikit-learn==1.3.2
spacy==3.7.2
numpy==1.24.4
pandas==2.1.4

# Database - PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Firebase
firebase-admin==6.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# API & Request Handling
python-multipart==0.0.6
requests==2.31.0

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.2
